[송원범] Controller 디버깅, 코드 방식 통일

+++ ComputerShoppingMall/src/main/java/controller/DeleteCoolerController.java
@@ -27,18 +27,28 @@ public class DeleteCoolerController extends HttpServlet {
 	}
 
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-	// 변수 등록
-	int coolerNo = 0;
-	// request값 받아오기
-	if(request.getParameter("coolerNo")!= null && request.getParameter("coolerNo") != "") {
-			coolerNo = Integer.parseInt(request.getParameter("coolerNo"));
+		// 세션확인
+		HttpSession session = request.getSession();
+		if((String)session.getAttribute("sessionAdminId") == null) {
+			// 로그인이 되어있지 않은 상태 -> 로그인 폼으로 돌아가기
+			response.sendRedirect(request.getContextPath() + "/LoginController");
+			return;
 		}
-	// 디버깅	
-	System.out.println(coolerNo+"<-coolerNo");
-	
-	// dao
-	coolerDao = new CoolerDao();
-	coolerDao.deleteCooler(coolerNo);
+		
+		// 변수 등록
+		int coolerNo = 0;
+		
+		// request값 받아오기
+		if(request.getParameter("coolerNo")!= null && request.getParameter("coolerNo") != "") {
+				coolerNo = Integer.parseInt(request.getParameter("coolerNo"));
+			}
+		
+		// 디버깅	
+		System.out.println(coolerNo+"<-coolerNo");
+		
+		// dao
+		coolerDao = new CoolerDao();
+		coolerDao.deleteCooler(coolerNo);
 
 		response.sendRedirect(request.getContextPath()+"/DigitalDownloadController");
 	}
